+++ acceptance-test/code-generator/petstore.yaml
