+++ acceptance-test/custom-code-generator/petstore.yaml
