+++ acceptance-test/custom-code-generator/src/main/groovy/MySpringCodegen.groovy
