+++ acceptance-test/custom-code-generator/src/template/server/api.mustache
