+++ acceptance-test/custom-code-generator/src/template/server/apiController.mustache
