+++ acceptance-test/custom-code-generator/src/template/server/apiException.mustache
