+++ acceptance-test/custom-code-generator/src/template/server/apiOriginFilter.mustache
