+++ acceptance-test/custom-code-generator/src/template/server/apiResponseMessage.mustache
