+++ acceptance-test/custom-code-generator/src/template/server/application.mustache
