+++ acceptance-test/custom-code-generator/src/template/server/enumClass.mustache
