+++ acceptance-test/custom-code-generator/src/template/server/enumOuterClass.mustache
