+++ acceptance-test/custom-code-generator/src/template/server/generatedAnnotation.mustache
