+++ acceptance-test/custom-code-generator/src/template/server/libraries/spring-boot/homeController.mustache
