+++ acceptance-test/custom-code-generator/src/template/server/libraries/spring-boot/pom.mustache
