+++ acceptance-test/custom-code-generator/src/template/server/libraries/spring-cloud/clientConfiguration.mustache
