+++ acceptance-test/custom-code-generator/src/template/server/libraries/spring-mvc/swaggerUiConfiguration.mustache
