+++ acceptance-test/custom-code-generator/src/template/server/libraries/spring-mvc/webApplication.mustache
