+++ acceptance-test/custom-code-generator/src/template/server/model.mustache
