+++ acceptance-test/custom-code-generator/src/template/server/notFoundException.mustache
