+++ acceptance-test/custom-code-generator/src/template/server/pathParams.mustache
