+++ acceptance-test/custom-code-generator/src/template/server/pojo.mustache
