+++ acceptance-test/custom-code-generator/src/template/server/project/plugins.sbt
