+++ acceptance-test/custom-code-generator/src/template/server/returnTypes.mustache
