+++ acceptance-test/custom-code-generator/src/template/server/swaggerDocumentationConfig.mustache
